ci: Enable a subset of the unit tests (#9)
### Motivation

The smoke tests need a registry as a test fixture.   We don't yet have that but until we do we can still run the other tests.

### Modifications

Enable the `unit-test` shared CI workflow, with a filter.

### Result

A subset of tests will run on each pull request.   This will ensure that the code builds and some functionality is exercised for each pull request.

### Test Plan

Ran the tests locally with `act pull_request`.