ContainerRegistry: Switch to swift-http-types (#11)
### Motivation

Swift HTTP Types provides currency types for HTTP clients and servers, which can be used with URLSession and AsyncHTTPClient.

Generally speaking, we try to keep Swift Container Plugin's dependencies to a minimum because each dependency increases the initial build time and also increases the chance of a version conflict with the project which is trying to use the plugin.   However `swift-http-types` is quite a small package and offers some nice ergonomic improvements over `URLRequest`/`URLResponse`, such as enums for HTTP status codes.

### Modifications

Switch all uses of `URLRequest`/`URLResponse` to `HTTPRequest`/`HTTPResponse`

### Result

No functional change.

### Test Plan

The unit and integration tests continue to pass.