ci: Enable smoke tests (#13)
### Motivation

The integration tests (`SmokeTests`) are not currently run on PRs because they require an instance of the container registry to be available.

### Modifications

Add a custom job which runs `registry` as a service container.

### Result

Integration tests will be run for every pull request.   The integration test job also runs the tests run by the unit tests job, but the duplication is not very costly. 
### Test Plan

All tests continue to pass.