Examples: Update Package.resolved for Hummingbird (#18)
### Motivation

Update Package.resolved, to pick up an important fix in swift-async-algorithms 1.0.2 which may prevent a crash in Hummingbird.

### Modifications

Update Package.resolved in the Hummingbird example.

### Result

The example will use newer dependencies.

### Test Plan

The crash is intermittent but seems to happen after the server has been idle.   A server built with the new pins has been idle without crashing for much longer than a server built with the old pins was able to run.