ContainerRegistry: Pass through the manifest's mediaType, if set (#23)
### Motivation

If the manifest blob's `mediaType` is set, the distribution spec requires that the HTTP `Content-Type` used to upload the manifest should match it.

### Modifications

In `putManifest`, use the blob's `mediaType` if it is set;  otherwise default to `manifest.v1`.

### Result

If a manifest uses a different media type, it will be passed through to the registry.   This might happen when copying a base image, which could have been created using the older manifest type.

### Test Plan

Automated tests continue to pass.  Also tested manually with different registries.