ContainerRegistry: Tolerate missing Content-Type header in registry response (#26)
### Motivation

ECR does not set the `Content-Type` header in its response to a blob upload.    If it is not present, the HTTP standard allows clients to guess the content type, or default to `application/octet-stream'.    In this case, knowing the content type is not critical so there is no need to check for it.

### Modifications

Do not check that responses contain the `Content-Type` header in the low-level HTTP client.

### Result

Blob uploads to ECR will not fail.

### Test Plan

Automated tests continue to pass; tested manually with ECR.