ContainerRegistry: Add a struct to represent a registry operation
Refactor RegistryClient, adding a struct which holds all the necessary
information about a requested operation on the registry.

This makes the separation of RegistryClient and HTTPClient's
responsibilities clearer, and allows us to centralise the generation
of registry URLs.  In the future it will make further improvements
easier, such as changing how authentication is handled and providing
more detailed error messages.