ContainerRegistry: Add a struct to represent a registry operation (#29)
### Motivation

`RegistryClient.executeRequestThrowing` currently takes an `HTTPRequest` struct parameter and executes it using the underlying `HTTPClient`.   It augments `HTTPClient` by handling authentication challenges and decodes JSON error responses, but the use of `HTTPRequest` means that the caller is responsible for details such as generating the full distribution endpoint URL.   Giving `RegistryClient` its own abstract operation type frees the caller from dealing with these internal details.

The struct can also be passed easily anywhere the details of the operation are required, making future improvements easier.   For instance, if an operation causes an error, the original operation struct could be attached to the error making all its details available to the error handler.

### Modifications

Add a struct which holds all the necessary information about a requested operation on the registry, and refactor `RegistryClient` to use it.

### Result

No functional change.

This refactoring makes the separation of RegistryClient and HTTPClient's responsibilities clearer, and allows us to centralise the generation of registry URLs.  In the future it will make further improvements easier, such as changing how authentication is handled and providing more detailed error messages.

### Test Plan

Automated tests continue to pass; also tested manually.