ContainerRegistry: Unconditionally depend on swift-crypto (#33)
### Motivation

`swift-crypto` provides a common cryptography interface on macOS and Linux.   On macOS it defers to CryptoKit, and although the files in the library must all be processed the compilation time is minimal.

In the past, on slower machines, avoiding processing `swift-crypto` on macOS seemed to improve build times by a few seconds.   Currently the cost of processing these files is negligible, and using `swift-crypto` on Linux and macOS allows conditional includes to be removed.

### Modifications

Always depend on `swift-crypto` in `Package.swift`
Replace conditional imports of `CryptoKit` with unconditional imports of `Crypto`.

### Result

No functional change.

### Test Plan

Automated tests continue to pass.