tests: Test main daily (#42)
Motivation
----------

It is very important that the example projects work when a new user
tries them, but upstream changes might cause them to stop working even
if nothing changes in this project. This PR adds a job to run the tests
daily to identify such problems as early as possible.

Modifications
-------------

* Extract the end to end tests into a reusable workflow
* Add a new action to run the tests daily, and for every push to main

Result
------

Tests will be run periodically.

Test Plan
---------

The existing tests continue to pass - this PR will run them more often.