tests: Mechanically convert to swift-testing (#45)
Motivation
----------

swift-testing offers a more convenient way to write test cases.   For
instance, it makes table-driven tests much easier to write than XCTest.

Modifications
-------------

This commit rewrites the existing XCTest test cases to use
swift-testing. It is a mechanical translation using the rules in:

   https://developer.apple.com/documentation/testing/migratingfromxctest

It does not try to adapt the style of the tests to `swift-testing`.

Result
------

The tests run using `swift-testing` with minimal changes. A future
change will adapt them to make better use of the new framework's
facilities.

One immediate improvement is that the `swift-testing` tests can be run
with the OSS Swift toolchain on macOS. `XCTests` cannot, because the
framework is not included in the OSS toolchain.

Test Plan
---------

All tests continue to pass.