tests: Convert ReferenceTests to `swift-testing` parameterised tests (#47)
Motivation
----------

`ImageReference.testReferences` tests is a table-driven test which
iterates over a list of test cases. Under `XCTest` it was implemented as
a single test containing a loop, but with `swift-testing` the loop can
be moved outside the test. This makes it easier to see that multiple
test cases are being run and to diagnose problems.

Modifications
-------------

`ImageReference.testReferences` now uses `@Test(arguments: ...)` to
iterate over the existing test list.

Result
------

Test reports are more detailed. 

Test Plan
---------

All tests continue to pass.