tests: Add a test for app layer ordering
`containertool` currently prepends the app layer to the layer stack,
but it should append it.   This means that a later layer will not
override a file with the same path in an earlier layer.

This commit adds an integration tests which stacks two layers on
top of each other using containertool and verifies that the second
layer overrides the first.