tests: Build the layering test images on scratch to avoid pulling swift:slim (#60)
Motivation
----------

The layering tests sometimes fail with 400 errors from the registry.
This error could be caused by pulling the underlying base image from
Docker Hub, although the logs do not give enough information to be
certain.

We don't need to run the container image in order to check that the
payload file is being layered correctly. This would avoid the need to
pull `swift:slim`, reducing the risk of external systems causing test
failures.

Modifications
-------------

Build the test image on scratch.
Instead of running the container and using `cat` to check the contents
of `/payload`, create a container from the image and copy `/payload`
out, without starting the container.
 
Result
------

The layering tests should be less susceptible to failures caused by
external systems.

Test Plan
---------

All tests continue to pass.