tar: Disallow empty filenames (#66)
Motivation
----------

A tar archive member name cannot be empty because a Unix filename cannot
be the empty string:


https://pubs.opengroup.org/onlinepubs/9699919799/basedefs/V1_chap03.html#tag_03_170

There would be no way to restore such a file to the filesystem.

Modifications
-------------

Throw an error if the archive member name is the empty string.

Result
------

Invalid archives containing empty filenames cannot be created.

Test Plan
---------

All existing tests pass.
A new test checks that an error is thrown if the member name is empty.