Examples: Add an example which uses a resource bundle (#82)
Motivation
----------

An example showing how to use resources (#78) is helpful for new users
and can be used as an end to end tests for this feature.

Modifications
-------------

* Add a new `HelloWorldWithResources` example
* Add the new example to the `End to end tests` GitHub Actions job
* Restructure the end to end tests to check the HTTP response code
instead of the content which is returned, because the new test returns
one of 3 JPEGs selected at random, whereas the existing tests returned
strings.
* Adopt the code structure used by the [official Hummingbird
examples](https://github.com/hummingbird-project/hummingbird-examples),
to make it more familiar to users who are used to the official examples.
 
Result
------

* The repository contains a new example showing how to use resources in
a container image
* The resource feature is tested as part of the automated end to end
test suite.

Test Plan
---------

The new test passes and all existing tests continue to pass.