docs: Split ContainerBuildPlugin documentation into multiple pages (#99)
Motivation
----------

Most of the documentation explaining how a user should run the plugin is
in a single DoCC file under the `ContainerImageBuilderPlugin` target
(this otherwise empty target is needed because DoCC does not generate
documentation for plugins).

This documentation would be easier to read, and easier to expand with
more information, if it was split into several sections. This would
provide an outline view in the left-hand pane (the "curation") which
would allow users to jump directly to sections of interest, and make
internal links between documentation sections easier.

Modifications
-------------

* Rename `ContainerImageBuilderPlugin` to `swift-container-plugin`.
`swift-container-plugin` will hold all the high level documentation,
examples and tutorials for the project. A future PR will add a manual
page which specifically covers only the plugin.
*
 
Result
------

The existing documentation is better structured. A number of small edits
have been made for clarity.

Test Plan
---------

* No functional change.   
* DoCC checks are already run by the `soundness` GitHub CI job and
continue to pass; all notes and usage warnings fixed.

---------

Co-authored-by: Joseph Heck <j_heck@apple.com>