docs: Add examples of building and publishing images (#108)
Motivation
----------

As with authentication configuration, it is useful for new users to be
able to refer to concrete examples showing how to use the plugin.

Modifications
-------------

* Add a section to the 'run' document showing how to build and run an
image in one command.
* Add sections showing how to publish images to some popular public
container registries.

Result
------

Users can refer to examples showing how to use the plugin to build and
run images.

Test Plan
---------

Steps tested manually.
All automated tests continue to pass.

---------

Co-authored-by: Joseph Heck <j_heck@apple.com>