docs: Point to main branch documentation on Swift Package Index (#112)
Motivation
----------

By default, Swift Package Index shows the documentation for the latest
release. This still has the `containertool` documentation target in
`.spi.yml`, which is causing the left-hand navigation bar (the curation)
to show the topics for `containertool` even though the main content is
the Swift Container Plugin Documentation.

This should be resolved when the next release is tagged, but in the
meantime linking to the `main` branch documentation will cause users to
see the most up-to-date documentation with the correct navigation bar.

Modifications
-------------

Links to the SPI documentation in `README.md` point to the `main`
branch.

Result
------

Users should see the latest documentation, with the correct navigation
bar.

Test Plan
---------

No functional change.
All existing tests continue to pass.