containertool: Read default username and password from the environment
Add environment variable equivalents for the `--username` and
`--password` command line flags.  These variables are more convenient
than .netrc when working with registries which use short-lived
credentials, such as ECR:

    export CONTAINERTOOL_USERNAME=AWS
    export CONTAINERTOOL_PASSWORD=$(aws ecr get-login-password --region us-west-2)w0
    swift run containertool --repository \
        123456789012.dkr.ecr.us-west-2.amazonaws.com/hello/world \
        .build/x86_64-swift-linux-musl/debug/hello-world

N.B. These flags are used as default credentials if no matching
entries are found in .netrc - potentially they should be called
`--default-username` and `--default-password` to make this clearer.

In the example above, if .netrc contains credentials for
123456789012.dkr.ecr.us-west-2.amazonaws.com they will be used in
preference to the credentials in the environment variables.
To avoid this, credentials which are not intended to be used should
be removed from .netrc.

Fixes #105