ContainerRegistry: Remove return-type overloaded client functions (#145)
Motivation
----------

Return type overloads can be puzzling for new readers because the same
function call returns a different type depending on where it is called.
In some cases a type annotation is needed at the call site, but when the
overloaded function is the final call in another function the compiler
can infer the type, so the reader needs to know to check the type of the
enclosing function.

In addition, in a future pull request we will need access to the
underlying data returned by the HTTP client, in order to verify digests
and create `ContentDescriptor` objects. This refactoring starts to
prepare for that and makes the code easier to understand.

Modifications
-------------

* Remove return-type overloaded version of `executeRequestThrowing()`
* Remove return-type overloaded version of `getBlob()`

Result
------

Refactoring.   No functional change.

Test Plan
---------

Existing tests continue to pass.