ContainerRegistry: Provide ContentDescriptors for manifests (#146)
Motivation
----------

In a multi-arch image, several architecture-specific manifests are
grouped together under the same
[index](https://github.com/opencontainers/image-spec/blob/v1.0.1/image-index.md).
The index has a list of [content
descriptors](https://github.com/opencontainers/image-spec/blob/v1.0.1/descriptor.md#properties)
pointing to the manifests.

Currently, `containertool` creates single-architecture images for which
the root object is a manifest. Returning a content descriptor for the
manifest allows it to create an index which points to the manifest.

On its own this does not allow `containertool` to create multi-arch
images, but it opens that possibility for the future. It is a necessary
towards being able to write [container images to
disk](https://github.com/opencontainers/image-spec/blob/v1.0.1/image-layout.md).

Modifications
-------------

* Return a ContentDescriptor from getManifest(), in addition to the
manifest object
* Return a ContentDescriptor from putManifest()

Result
------

`containertool` can retrieve a `ContentDescriptor` for a manifest it has
written or fetched.

Test Plan
---------

Existing tests have been updated to match the new signatures and
continue to pass.