ContainerRegistry: Make digest(of:Data) an ImageReference.Digest constructor (#147)
Motivation
----------

The `digest(of:)` function calculates an image digest from a `Data`
object. Previously the digest was a string and the function stood alone,
but after the introduction of the `ImageReference.Digest` type it is
logical for it to be an alternative constructor.

Modifications
-------------

`digest(of:)` is now `ImageReference.Digest(of:)`

Result
------

Refactoring.   No functional change.

Test Plan
---------

All existing tests continue to pass.