ContainerRegistry: Separate registry operations into `ImageSource` and `ImageDestination` protocols (#148)
Motivation
----------

Currently, `RegistryClient` handles uploading and downloading container
images to and from a registry. In fact, `containertool` creates two
clients: one is used only to download images from the source repository,
and the other is used only to upload images to the destination.
Registries are read/write, but in the future we may need to handle
sources of images which are read-only or write-only - for example when
writing a container image out to an archive on disk.

This change adds two new protocols, splitting up the operations needed
to upload and download images. This allows for future clients which
cannot handle both sets of operations.

Modifications
-------------

* Add new `ImageSource` and `ImageDestination` protocols, defining the
functions needed to download and upload images, respectively.
* Consolidate the `RegistryClient` extensions which implement the
`ImageSource` and `ImageDestination` protocols.

Result
------

Refactoring. No functional change, but responsibilities of different
objects are clearer.

Test Plan
---------

Existing tests continue to pass.