containertool: Store an index, pointing to the image manifest (#151)
Motivation
----------

A multiarch image consists of an index pointing to one or more
architecture-specific manifests. `containertool` currently builds a
single architecture image, which can be pushed using just a manifest,
but including the index as well is a step towards being able to build
multiarch images in future.

Modifications
-------------

* Add `putIndex` operation to the `ImageDestination` protocol and
implement it in `RegistryClient`
* Build an index in `publishContainerImage` and print its reference on
standard output, instead of the reference of the manifest.

Result
------

Although it still builds single-architecture images, a `containertool`
image will start with an index pointing to a manifest, instead of a bare
manifest.

Test Plan
---------

All existing tests continue to pass.