Update Package.swift (#153)
Motivation
----------

I _believe_ the line `.package(path: "../.."),` in the `Package.swift`
of the Hummingbird example is vestigial from perhaps a previous version
where it was just a `Package.swift` and a `main.swift`. If it is need
explicitly for a container build, please mention in the docs because a
regular `swift build` to test the app locally doesn't work until it is
removed.

Modifications
-------------

removed the single line `.package(path: "../..")` from the
`Package.swift` in the Hummingbird example.

Result
------

Example will now build to be tested pre-containerization test. 

Test Plan
---------

Don't believe change needs a test as the example already does not have
tests.

---------

Co-authored-by: Euan Harris <euanh@apple.com>